feat: Add groups field to User model
The code changes in this commit add a new field `groups` to the User model in the `User.js` file. This field is an array of `Group` objects and is used to store the groups that the user belongs to.

This change allows for better organization and management of user groups within the application. Users can now be associated with multiple groups, providing more flexibility and functionality.

Based on the recent user commits and repository commits, the commit message follows the established convention of starting with a verb in the imperative form (`Add`) and providing a concise description of the changes (`groups field to User model`).